Add state managers to WebApi and AdminWebApi
